% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{.valid_enquos}
\alias{.valid_enquos}
\title{Keep only valid quosures (not missing, not NULL)}
\usage{
.valid_enquos(quos)
}
\arguments{
\item{quos}{A list of quosures (from \code{\link[rlang:enquo]{rlang::enquos()}})}
}
\value{
A filtered list of quosures
}
\description{
Utility to filter quosures returned by \code{\link[rlang:enquo]{rlang::enquos()}},
dropping those that are missing or explicitly NULL.
}
\keyword{internal}
