% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_hline1}
\alias{geom_hline1}
\title{Horizontal line 1}
\usage{
geom_hline1(
  logscale = FALSE,
  yintercept = 1,
  color = "red",
  linetype = "dashed",
  ...
)
}
\arguments{
\item{logscale}{a logical whether to apply log scale.}

\item{yintercept}{a numeric that control the position of the horizontal line.}

\item{color}{a character specifying a color.}

\item{linetype}{\link{linetype}}

\item{...}{other arguments to pass to \code{\link[ggplot2]{geom_hline}}.}
}
\value{
a ggproto object
}
\description{
Draw a horizontal line 1.
}
