% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_legend}
\alias{get_legend}
\title{Extract the legend from a ggplot}
\usage{
get_legend(plot)
}
\arguments{
\item{plot}{A ggplot object.}
}
\value{
A legend grob (the \code{"guide-box"} grob) extracted from plot.
}
\description{
Get the legend grob from a ggplot object.
}
\examples{
\donttest{
set.seed(123)
dat <- expand.grid(x = c("A","B","C"), fill = c("X","Y","Z"))
dat$y <- sample(1:10, size = 9, replace = TRUE)
g <- ggbar(dat, x = x, y = y, fill = fill, label = y,
           label_args = list(vjust = -0.25)) +
     theme_view(family = NULL)
get_legend(g)
}

}
