% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{ggpie}
\alias{ggpie}
\title{ggplot pie functions only with frequently used arguments}
\usage{
ggpie(
  data,
  y,
  group,
  text,
  label,
  label_family = "Comic Sans MS",
  label_size = 4,
  label_angle = 0,
  label_hjust = 0.5,
  label_vjust = 0.5,
  label_color = c("#000000", "#FAF9F6")
)
}
\arguments{
\item{data}{A data.frame}

\item{y}{A name of axis \code{y}}

\item{group}{A name of variable you want to group}

\item{text}{A name of variable or expression for ggplotly hover text}

\item{label}{A name of variable or expression you want to label}

\item{label_family, label_size, label_angle}{A string specifying label font-family, size and angle}

\item{label_hjust, label_vjust}{A numeric specifying label horizontal and vertical
adjustment}

\item{label_color}{A string specifying label color}
}
\value{
A ggplot object
}
\description{
Shortly simplify the grammar of ggplot to the functions only with frequently
used arguments. (\code{ggbar}, \code{ggline}, \code{ggpoint}, \code{ggjitter}, \code{ggmix}, \code{ggpie}, \code{ggtable})
}
\examples{
# pie
\donttest{set.seed(123)
data <- data.frame(y = c(60, 30, 10), group = c("A", "B", "C"))
ggpie(data = data, y = y, group = group, label = sprintf("\%s\%\%", y),
      label_family = NA)}

}
\seealso{
\code{\link[=ggbar]{ggbar()}}, \code{\link[=ggline]{ggline()}}, \code{\link[=ggpoint]{ggpoint()}}, \code{\link[=ggjitter]{ggjitter()}}, \code{\link[=ggmix]{ggmix()}}, \code{\link[=ggtable]{ggtable()}}
}
