% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{ggpoint}
\alias{ggpoint}
\alias{ggjitter}
\title{ggplot point functions only with frequently used arguments}
\usage{
ggpoint(
  data,
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  group = NULL,
  color = NULL,
  fill = NULL,
  shape = NULL,
  size = NULL,
  text = NULL,
  label,
  label_family = "Comic Sans MS",
  label_size = 4,
  label_angle = 0,
  label_hjust = 0.5,
  label_vjust = 0.5,
  label_color = c("#000000", "#FAF9F6")
)

ggjitter(
  data,
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  group = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  text = NULL,
  label,
  label_family = "Comic Sans MS",
  label_size = 4,
  label_angle = 0,
  label_hjust = 0.5,
  label_vjust = 0.5,
  label_color = c("#000000", "#FAF9F6")
)
}
\arguments{
\item{data}{A data.frame}

\item{x, y}{A name of axis \code{x} and \code{y}}

\item{ymin, ymax}{\code{min} and \code{max} values of y for the height}

\item{group, color, fill}{A name of variable you want to group, color and fill}

\item{shape}{A name of point shape variable}

\item{size}{A name of point size variable}

\item{text}{A name of variable or expression for ggplotly hover text}

\item{label}{A name of variable or expression you want to label}

\item{label_family, label_size, label_angle}{A string specifying label font-family, size and angle}

\item{label_hjust, label_vjust}{A numeric specifying label horizontal and vertical
adjustment}

\item{label_color}{A string specifying label color}
}
\value{
A ggplot object
}
\description{
Shortly simplify the grammar of ggplot to the functions only with frequently
used arguments.
}
\examples{
# point
\donttest{data <- expand.grid(x = 1:10, y = 1:10)
data$shape <- sample(x = c("A", "B", "C"), size = 10, replace = TRUE)
ggpoint(data = data, x = x, y = y, shape = shape, label = y, label_family = NA) +
  theme_view(family = NA)
 ggjitter(data = data, x = x, y = y, shape = shape, size = y, label = shape,
          color = shape, label_family = NA) +
  theme_view(family = NULL)
}

}
\seealso{
\code{\link[=ggbar]{ggbar()}}, \code{\link[=ggline]{ggline()}}, \code{\link[=ggjitter]{ggjitter()}}, \code{\link[=ggmix]{ggmix()}}, \code{\link[=ggpie]{ggpie()}}, \code{\link[=ggtable]{ggtable()}}
}
