% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshort-package.R
\docType{package}
\name{ggshort-package}
\alias{ggshort}
\alias{ggshort-package}
\title{ggshort: ggplot shortly}
\description{
This package simplify the grammar of ggplot to the functions only with
frequently used arguments. Despite its limited flexibility, it helps people
draw ggplot easily.
}
\author{
\strong{Maintainer}: Seokhoon Last \email{seokhoonj@gmail.com}

}
\keyword{internal}
