% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grob2ggplot}
\alias{grob2ggplot}
\title{Grob to ggplot}
\usage{
grob2ggplot(plot)
}
\arguments{
\item{plot}{a grob object}
}
\value{
a ggplot object.
}
\description{
Change a grob object to a ggplot object.
}
