% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grob_to_ggplot}
\alias{grob_to_ggplot}
\title{Convert grob-like objects to a ggplot}
\usage{
grob_to_ggplot(grob)
}
\arguments{
\item{grob}{A ggplot, gtable or grob.}
}
\value{
A ggplot object displaying the input.
}
\description{
Wrap a grid grob or gtable inside an empty ggplot so it can be displayed
or combined consistently with other ggplots.
}
\details{
\itemize{
\item If \code{grob} is already a ggplot, it is returned unchanged.
\item If \code{grob} is a gtable or grob, it is embedded using
\code{\link[ggplot2:annotation_custom]{ggplot2::annotation_custom()}}.
}
}
\examples{
\donttest{
p <- ggplot2::qplot(mpg, hp, data = mtcars)
grob <- ggplot2::ggplotGrob(p)
grob_to_ggplot(grob) # convert gtable back to ggplot
}

}
