% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{match_theme}
\alias{match_theme}
\title{Match theme}
\usage{
match_theme(theme = c("view", "save", "shiny"), ...)
}
\arguments{
\item{theme}{a string specifying a ggshort theme function ("view", "save", "shiny")}

\item{...}{arguments passed on to (\code{theme_view}, \code{theme_save}, \code{theme_shiny})}
}
\description{
Match a ggshort theme.
}
