% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{metaplot}
\alias{metaplot}
\title{Meta plot}
\usage{
metaplot(x, widths = c(5, 5), theme = c("view", "save", "shiny"))
}
\arguments{
\item{x}{a meta object}

\item{widths}{a unit vector giving the width of each column}

\item{theme}{a string specifying a ggshort theme function ("view", "save", "shiny")}
}
\value{
a ggplot object
}
\description{
Draw a plot for a meta object.
}
\examples{
# meta plot
\dontrun{
mt <- jaid::meta(mtcars)
metaplot(mt)}

}
