% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaplot}
\alias{pcaplot}
\title{PCA plot}
\usage{
pcaplot(
  object,
  x = 1,
  y = 2,
  target = NULL,
  scale = 1,
  alpha = 0.3,
  label_family = "Comic Sans MS",
  theme = c("view", "save", "shiny"),
  ...
)
}
\arguments{
\item{object}{a prcomp object}

\item{x}{a priciple component rank for a x axis (default: 1)}

\item{y}{a priciple component rank for a y axis (default: 2)}

\item{target}{a vector for grouping points (default: NULL)}

\item{scale}{a scaling parameter, disabled by 0 (default: 1)}

\item{alpha}{a point color transparency}

\item{label_family}{a string specifying label font-family}

\item{theme}{a string specifying a ggshort theme function ("view", "save", "shiny")}
}
\value{
a ggplot object
}
\description{
Draw a PCA plot for a prcomp object.
}
\examples{
# PCA plot
\dontrun{
pca <- prcomp(iris[, -5])
pcaplot(pca, x = 1, y = 1, target = iris$Species)}

}
