% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_fill_by_month_gradientn}
\alias{scale_fill_by_month_gradientn}
\title{Continuous fill gradient by month (for Date variables)}
\usage{
scale_fill_by_month_gradientn(
  by_month = 6,
  palette = function(n) grDevices::hcl.colors(n, "YlGnBu"),
  n = 256,
  include_endpoints = FALSE,
  ...
)
}
\arguments{
\item{by_month}{Integer; interval in months between colorbar ticks.
Default is \code{6}.}

\item{palette}{Color palette function or vector. Default is
\code{\link[grDevices:palettes]{grDevices::hcl.colors()}} with "YlGnBu" palette.}

\item{n}{Integer; number of colors sampled from the palette. Default \code{256}.}

\item{include_endpoints}{Logical; include the min/max dates as ticks if \code{TRUE}.
Default \code{FALSE} (to avoid label overlap).}

\item{...}{Additional args passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}.}
}
\value{
A ggplot2 fill scale for month-based Date aesthetics.
}
\description{
A ggplot2 continuous fill scale for Date variables.
Generates colorbar breaks every N months (default: 6 months),
labeled as "YYYY-MM".
}
