% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_x_comma}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_comma}
\title{Add commas to x-axis or y-axis tick labels}
\usage{
scale_x_comma()

scale_y_comma()

scale_comma()
}
\value{
an ggproto object
}
\description{
Add commas to x-axis or y-axis tick labels.
}
