% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_x_limit_reverse}
\alias{scale_x_limit_reverse}
\alias{scale_y_limit_reverse}
\title{Scale limit reverse (x & y)}
\usage{
scale_x_limit_reverse(x)

scale_y_limit_reverse(x)
}
\arguments{
\item{x}{a vector x or y variable}
}
\value{
a ggproto object
}
\description{
Reverse x & y axis scale limit
}
