% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_x_log}
\alias{scale_x_log}
\alias{scale_y_log}
\title{Logarithmic scale with custom base}
\usage{
scale_x_log(base = exp(1), ...)

scale_y_log(base = exp(1), ...)
}
\arguments{
\item{base}{Logarithm base. Default is \code{exp(1)} (natural log).}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}
or \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Shortcut for applying a logarithmic transformation to the x- or y-axis.
Defaults to the natural log (\code{base = exp(1)}).
}
\examples{
\dontrun{
ggpoint(mtcars, mpg, wt) +
  scale_x_log()                 # natural log on x

ggpoint(mtcars, mpg, wt) +
  scale_y_log(base = 10)        # log10 on y

ggpoint(mtcars, mpg, wt) +
  scale_x_log(base = 2)         # log base 2 on x
}

}
