% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_x_log1p}
\alias{scale_x_log1p}
\alias{scale_y_log1p}
\title{Log(1 + x) scale (safe logarithmic transform)}
\usage{
scale_x_log1p(...)

scale_y_log1p(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}
or \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}
}
\description{
Shortcut for applying a \code{log1p()} (i.e., \code{log(1 + x)}) transformation
to the x- or y-axis. This is numerically stable for small or zero values.
}
\examples{
\dontrun{
ggpoint(mtcars, mpg, wt) +
  scale_x_log1p() +
  scale_y_log1p()
}

}
