% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_density_quantile_vline}
\alias{stat_density_quantile_vline}
\title{Vertical quantile line(s) for density plots}
\usage{
stat_density_quantile_vline(
  mapping = NULL,
  data = NULL,
  probs = 0.95,
  na.rm = FALSE,
  linetype = "dashed",
  linewidth = 0.3,
  show.legend = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{probs}{Numeric vector of probabilities in [0, 1] (e.g., \code{0.95}).}

\item{na.rm}{Logical; remove NAs silently if \code{TRUE}. Default \code{FALSE}.}

\item{linetype, linewidth}{Line style/width for the vline(s).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:geom_abline]{ggplot2::geom_vline()}}.}
}
\value{
A ggplot layer.
}
\description{
Draw vertical line(s) at quantile(s) of \code{x}. Respects grouping if a
grouping/color/fill aesthetic is mapped.
}
\examples{
\donttest{
ggdensity(iris, x = Sepal.Length, fill = Species) +
  stat_density_quantile_vline(probs = c(.5, .9))
}

}
