% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{switch_theme}
\alias{switch_theme}
\title{Switch theme}
\usage{
switch_theme(theme = c("view", "save", "shiny"), ...)
}
\arguments{
\item{theme}{One of \code{"view"}, \code{"save"}, \code{"shiny"}.}

\item{...}{Additional arguments passed to the chosen theme function.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} object.
}
\description{
Switch between built-in ggshort themes: \code{"view"}, \code{"save"}, \code{"shiny"}.
}
