% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{corplot}
\alias{corplot}
\title{Correlation plot}
\usage{
corplot(
  x,
  display = c("full", "lower", "upper"),
  theme = c("view", "save", "shiny")
)
}
\arguments{
\item{x}{a correlation matrix}

\item{display}{a matrix display type ("full", "lower", "upper")}

\item{theme}{a string specifying a ggshort theme function ("view", "save", "shiny")}
}
\value{
a ggplot object
}
\description{
Draw a plot for a correlation matrix.
}
\examples{
# correlation plot
\dontrun{
x <- cor(mtcars)
corplot(x)}

}
