% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_legend}
\alias{get_legend}
\title{Get a legend}
\usage{
get_legend(plot)
}
\arguments{
\item{plot}{a ggplot object}
}
\value{
a ggplot object
}
\description{
Get a legend object from the ggplot object.
}
\examples{
# get a legend.
\donttest{set.seed(123)
data <- expand.grid(x = c("A", "B", "C"), fill = c("X", "Y", "Z"))
data$y <- sample(x = 1:10, size = 9, replace = TRUE)
g <- ggbar(data = data, x = x, y = y, fill = fill, label = y, label_vjust = -.25) +
  theme_view(family = NULL)
get_legend(g)}

}
