% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{ggbar}
\alias{ggbar}
\title{ggplot bar helper (frequently used arguments)}
\usage{
ggbar(
  data,
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  ymin_err,
  ymax_err,
  color = NULL,
  fill = NULL,
  group = NULL,
  text = NULL,
  label,
  label_args = list(family = getOption("ggshort.font"), size = 4, angle = 0, hjust = 0.5,
    vjust = 0.5, color = "black")
)
}
\arguments{
\item{data}{A data.frame.}

\item{x, y}{Unquoted column names for x and y aesthetics.}

\item{ymin, ymax}{Optional lower/upper bounds for y (rarely used for bars).}

\item{ymin_err, ymax_err}{Optional lower/upper bounds for error bars.}

\item{color, fill, group}{Optional columns mapped to the \code{color}, \code{fill} and
\code{group} aesthetics.}

\item{text}{Optional column/expression for tooltip text (e.g., ggplotly).}

\item{label}{Unquoted column/expression to use as text label.}

\item{label_args}{A named list of text style options for \code{geom_text()}.
Supported names: \code{family}, \code{size}, \code{angle}, \code{hjust}, \code{vjust}, \code{color}.}
}
\value{
A ggplot object.
}
\description{
Quickly build bar charts with common options. Supply \strong{unquoted}
column names.
}
\examples{
\donttest{
set.seed(123)
df <- expand.grid(x = c("A","B","C"), fill = c("X","Y","Z"))
n <- nrow(df); df$y <- sample(10:30, n, TRUE)
se <- runif(n, 1.5, 4)
df$ymin_err <- pmax(df$y - se, 0); df$ymax_err <- df$y + se
ggbar(df, x = x, y = y, fill = fill,
      ymin_err = ymin_err, ymax_err = ymax_err,
      label = y, label_args = list(vjust = -0.25)) +
  theme_view()
}

}
\seealso{
\code{\link[=ggline]{ggline()}}, \code{\link[=ggpoint]{ggpoint()}}, \code{\link[=ggjitter]{ggjitter()}}, \code{\link[=ggscatter]{ggscatter()}}, \code{\link[=ggdensity]{ggdensity()}},
\code{\link[=gghistogram]{gghistogram()}}, \code{\link[=ggbox]{ggbox()}}, \code{\link[=ggpie]{ggpie()}}, \code{\link[=ggmix]{ggmix()}}, \code{\link[=ggtable]{ggtable()}}
}
