% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{ggline}
\alias{ggline}
\title{ggplot line functions only with frequently used arguments}
\usage{
ggline(
  data,
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  ymin_err,
  ymax_err,
  group = NULL,
  color = NULL,
  linetype = NULL,
  text = NULL,
  label,
  label_family = "Comic Sans MS",
  label_size = 4,
  label_angle = 0,
  label_hjust = 0.5,
  label_vjust = 0.5,
  label_color = c("#000000", "#FAF9F6")
)
}
\arguments{
\item{data}{a data.frame}

\item{x, y}{a name of axis \code{x} and \code{y}}

\item{ymin, ymax}{\code{min} and \code{max} values of y for the height}

\item{ymin_err, ymax_err}{\code{min} and \code{max} values of y errors for the error bar}

\item{group, color}{a name of variable you want to group and color}

\item{linetype}{a name of linetype variable}

\item{text}{a name of variable or expression for ggplotly hover text}

\item{label}{a name of variable or expression you want to label}

\item{label_family, label_size, label_angle}{a string specifying label font-family, size and angle}

\item{label_hjust, label_vjust}{a numeric specifying label horizontal and vertical
adjustment}

\item{label_color}{a string specifying label color}
}
\value{
a ggplot object
}
\description{
Shortly simplify the grammar of ggplot to the functions only with frequently
used arguments.
}
\examples{
# line
\donttest{set.seed(123)
data <- expand.grid(x = 1:20, color = c("X", "Y", "Z"))
data$y <- sample(1:10, size = nrow(data), replace = TRUE)
ggline(data = data, x = x, y = y, color = color) +
  theme_view(family = NULL)}

}
\seealso{
\code{\link[=ggbar]{ggbar()}}, \code{\link[=ggpoint]{ggpoint()}}, \code{\link[=ggjitter]{ggjitter()}}, \code{\link[=ggmix]{ggmix()}}, \code{\link[=ggpie]{ggpie()}}, \code{\link[=ggtable]{ggtable()}}
}
