% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{ggmix}
\alias{ggmix}
\title{ggplot mix functions only with frequently used arguments}
\usage{
ggmix(
  data,
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  group = NULL,
  fill = NULL,
  bar_color = "transparent",
  text = NULL,
  label,
  label_family = "Comic Sans MS",
  label_size = 4,
  label_angle = 0,
  label_hjust = 0.5,
  label_vjust = 0.5,
  label_color = c("#000000", "#FAF9F6"),
  reverse = FALSE
)
}
\arguments{
\item{data}{a data.frame}

\item{x, y}{a name of axis \code{x} and \code{y}}

\item{ymin, ymax}{\code{min} and \code{max} values of y for the height}

\item{group, fill}{a name of variable you want to group and fill}

\item{bar_color}{a string specifying bar color}

\item{text}{a name of variable or expression for ggplotly hover text}

\item{label}{a name of variable or expression you want to label}

\item{label_family, label_size, label_angle}{a string specifying label font-family, size and angle}

\item{label_hjust, label_vjust}{a numeric specifying label horizontal and vertical
adjustment}

\item{label_color}{a string specifying label color}

\item{reverse}{a boolean whether to reverse the order of the \code{y} variable}
}
\value{
a ggplot object
}
\description{
Shortly simplify the grammar of ggplot to the functions only with frequently
used arguments. (\code{ggbar}, \code{ggline}, \code{ggpoint}, \code{ggjitter}, \code{ggmix}, \code{ggpie}, \code{ggtable})
}
\examples{
# mix
\donttest{set.seed(123)
data <- expand.grid(x = c("A", "B", "C"), fill = c("X", "Y", "Z"))
data$y <- sample(x = 1:10, size = 9, replace = TRUE)
ggmix(data = data, x = x, y = y, fill = fill, label = y, label_family = NA,
  reverse = TRUE) +
  theme_view(family = NULL)}

}
\seealso{
\code{\link[=ggbar]{ggbar()}}, \code{\link[=ggline]{ggline()}}, \code{\link[=ggpoint]{ggpoint()}}, \code{\link[=ggjitter]{ggjitter()}}, \code{\link[=ggpie]{ggpie()}}, \code{\link[=ggtable]{ggtable()}}
}
