% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{ggpie}
\alias{ggpie}
\title{ggplot pie functions only with frequently used arguments}
\usage{
ggpie(
  data,
  group,
  value,
  text,
  label,
  label_family = "Comic Sans MS",
  label_size = 4,
  label_angle = 0,
  label_hjust = 0.5,
  label_vjust = 0.5,
  label_color = c("#000000", "#FAF9F6")
)
}
\arguments{
\item{data}{a data.frame}

\item{group}{a name of variable to group}

\item{value}{a name of variable specifying values}

\item{text}{a name of variable or expression for ggplotly hover text}

\item{label}{a name of variable or expression to label}

\item{label_family, label_size, label_angle}{a string specifying label font-family, size and angle}

\item{label_hjust, label_vjust}{a numeric specifying label horizontal and vertical
adjustment}

\item{label_color}{a string specifying label color}
}
\value{
a ggplot object
}
\description{
Shortly simplify the grammar of ggplot to the functions only with frequently
used arguments. (\code{ggbar}, \code{ggline}, \code{ggpoint}, \code{ggjitter}, \code{ggscatter}, \code{ggmix}, \code{ggpie}, \code{ggtable})
}
\examples{
# pie
\donttest{set.seed(123)
data <- data.frame(group = c("A", "B", "C"), value = c(60, 30, 10))
ggpie(data = data, group = group, value = value, label = sprintf("\%s\%\%", value),
      label_family = NA)}

}
\seealso{
\code{\link[=ggbar]{ggbar()}}, \code{\link[=ggline]{ggline()}}, \code{\link[=ggpoint]{ggpoint()}}, \code{\link[=ggjitter]{ggjitter()}}, \code{\link[=ggscatter]{ggscatter()}}, \code{\link[=ggmix]{ggmix()}}, \code{\link[=ggtable]{ggtable()}}
}
