% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short.R
\name{ggtable}
\alias{ggtable}
\title{ggplot table functions only with frequently used arguments}
\usage{
ggtable(
  data,
  x,
  y,
  linetype = "solid",
  text = NULL,
  label,
  label_family = "Comic Sans MS",
  label_size = 4,
  label_angle = 0,
  label_hjust = 0.5,
  label_vjust = 0.5,
  label_color = c("#000000", "#FAF9F6"),
  xlab_position = c("bottom", "top")
)
}
\arguments{
\item{data}{a data.frame}

\item{x, y}{a name of axis \code{x} and \code{y}}

\item{linetype}{a string specifying a linetype (default: solid)}

\item{text}{a name of variable or expression for ggplotly hover text}

\item{label}{a name of variable or expression you want to label}

\item{label_family, label_size, label_angle}{a string specifying label font-family, size and angle}

\item{label_hjust, label_vjust}{a numeric specifying label horizontal and vertical
adjustment}

\item{label_color}{a string specifying label color}

\item{xlab_position}{a string specifying x label position (default: bottom)}
}
\value{
a ggplot object
}
\description{
Shortly simplify the grammar of ggplot to the functions only with frequently
used arguments. (\code{ggbar}, \code{ggline}, \code{ggpoint}, \code{ggjitter}, \code{ggscatter}, \code{ggmix}, \code{ggpie}, \code{ggtable})
}
\examples{
# table
\donttest{set.seed(123)
data <- expand.grid(x = c("A", "B", "C"), y = c("X", "Y", "Z"))
data$label <- sample(x = 1:10, size = 9, replace = TRUE)
ggtable(data = data, x = x, y = y, label = label, label_family = NA) +
  theme_view(family = NULL)}

}
\seealso{
\code{\link[=ggbar]{ggbar()}}, \code{\link[=ggline]{ggline()}}, \code{\link[=ggpoint]{ggpoint()}}, \code{\link[=ggjitter]{ggjitter()}}, \code{\link[=ggscatter]{ggscatter()}}, \code{\link[=ggmix]{ggmix()}}, \code{\link[=ggpie]{ggpie()}}
}
