% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{grid_top_bottom_right}
\alias{grid_top_bottom_right}
\alias{grid_left_to_right}
\title{Grid top plot, bottom plot and right legend}
\usage{
grid_top_bottom_right(g1, g2, legend, heights = c(5, 5), widths = c(8.5, 1.5))

grid_left_to_right(g1, g2, legend, widths = c(4, 4, 2))
}
\arguments{
\item{g1}{a ggplot object}

\item{g2}{a ggplot object}

\item{legend}{a gtable object}

\item{heights, widths}{a unit vector giving the height of each row or the widths of each column}
}
\value{
a gtable object
}
\description{
Grid top plot, bottom plot and right legend
}
\examples{
# grid top plot, bottom plot, right legend
\donttest{set.seed(123)
data <- expand.grid(x = c("A", "B", "C"), fill = c("X", "Y", "Z"))
data$y <- sample(x = 1:10, size = 9, replace = TRUE)
g1 <- ggbar(data = data, x = x, y = y, fill = fill, label = y, label_vjust = -.25,
  label_family = NA) +
  theme_view(family = NULL)
set.seed(123)
data <- expand.grid(x = 1:20, color = c("X", "Y", "Z"))
data$y <- sample(1:10, size = nrow(data), replace = TRUE)
g2 <- ggline(data = data, x = x, y = y, color = color) +
  theme_view(family = NULL)
legend <- get_legend(g1)
grid_top_bottom_right(g1, g2, legend)}

}
