% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{match_theme}
\alias{match_theme}
\title{Deprecated: match_theme()}
\usage{
match_theme(theme = c("view", "save", "shiny"), ...)
}
\arguments{
\item{theme}{One of \code{"view"}, \code{"save"}, \code{"shiny"}.}

\item{...}{Additional arguments passed to the chosen theme function.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Deprecated alias for \code{\link[=switch_theme]{switch_theme()}}.
Use \code{\link[=switch_theme]{switch_theme()}} instead.
}
