% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_pie}
\alias{plotly_pie}
\title{Plotly pie}
\usage{
plotly_pie(
  data,
  labels,
  values,
  texttemplate = "\%{label}<br>\%{value}<br>(\%{percent})",
  hovertemplate = "\%{label}<br>\%{value}<br>(\%{percent})<extra></extra>"
)
}
\arguments{
\item{data}{a grouping data frame}

\item{labels}{a name of variable you want to group}

\item{values}{a name of variable specifying values}

\item{texttemplate}{a string specifying a text template}

\item{hovertemplate}{a string specifying a hover template}
}
\description{
Draw a Plotly pie plot using grouping data.
}
\examples{
# plotly pie
\donttest{set.seed(123)
data <- data.frame(labels = c("A", "B", "C"), values = c(300, 150, 50))
plotly_pie(data = data, labels = labels, values = values)}

}
