% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{plotly_treemap}
\alias{plotly_treemap}
\title{Plotly treemap}
\usage{
plotly_treemap(
  data,
  group_var,
  value_var,
  path_labels = FALSE,
  path_sep = " / ",
  add_parents = FALSE,
  textinfo = "label+value+percent entry+percent parent+percent root",
  hoverinfo = "percent entry+percent parent+percent root"
)
}
\arguments{
\item{data}{a grouping data frame}

\item{group_var}{names of group variables}

\item{value_var}{a variable name to sum by group variables}

\item{path_labels}{a boolean value whether to use labels like path type}

\item{path_sep}{a string to separate the path labels. (default: " / ")}

\item{add_parents}{a boolean value whether the parents are prefixed to the labels}

\item{textinfo}{a string to express text information
(default: "label+value+percent entry+percent parent+percent root")}

\item{hoverinfo}{a string to express hover information
(default: "percent entry+percent parent+percent root")}
}
\value{
a plotly object
}
\description{
Draw a Plotly treemap using grouping data
}
\examples{
# plotly treemap
\dontrun{set.seed(123)
x <- c("A", "A", "B", "B", "C", "C")
y <- c("X", "Y", "X", "Y", "X", "Y")
z <- c(600, 500, 400, 300, 200, 100)
data <- data.frame(x = x, y = y, z = z)
plotly_treemap(data, group_var = .(x, y), value_var = z)}

}
