% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_pair_color_manual}
\alias{scale_pair_color_manual}
\alias{scale_pair_fill_manual}
\title{Create pair fill scales}
\usage{
scale_pair_color_manual(
  pair,
  pair_levels = c("1", "2"),
  color_type = c("base", "deep"),
  guide = "legend"
)

scale_pair_fill_manual(
  pair,
  pair_levels = c("1", "2"),
  color_type = c("base", "deep"),
  guide = "legend"
)
}
\arguments{
\item{pair}{A vector contains data of pair values}

\item{pair_levels}{Two element vector expressing pair values (default, c("1", "2"))}

\item{color_type}{A string of color type, \code{base} and \code{deep}}

\item{guide}{A function used to create a guide or its name. See \code{\link[=guides]{guides()}} for more information.}
}
\value{
A ggplot object
}
\description{
Create consistent pair fill scales
}
\examples{
# pair values with `ggline()`
\donttest{data <- expand.grid(agecat = c(1:10), gender = c("1", "2"))
data$n <- c(1:10, 2:11)
ggline(data, x = agecat, y = n, color = gender) +
  scale_pair_color_manual(data$gender) +
  theme_view(family = NULL)}

# pair values with `ggbar()`
\donttest{data <- expand.grid(agecat = c(1:10), gender = c("1", "2"))
data$n <- c(1:10, 2:11)
ggbar(data, x = agecat, y = n, fill = gender) +
  scale_pair_fill_manual(data$gender, color_type = "deep") +
  theme_view(family = NULL)}

}
\seealso{
\code{\link[=scale_pair_fill_manual]{scale_pair_fill_manual()}}
}
