% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_x_comma}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_comma}
\title{Add comma-formatted axis tick labels}
\usage{
scale_x_comma()

scale_y_comma()

scale_comma()
}
\value{
A ggproto scale object.

A list of two scales: \code{scale_x_comma()} and \code{scale_y_comma()}.
}
\description{
Convenience scales that format numeric axis labels with commas.
}
\seealso{
\code{\link[scales:comma]{scales::comma()}}, \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}, \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}
}
