% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_view}
\alias{theme_view}
\title{View theme}
\usage{
theme_view(
  family = "Comic Sans MS",
  x.size = NULL,
  y.size = NULL,
  t.size = NULL,
  s.size = NULL,
  l.size = NULL,
  x.face = "plain",
  y.face = "plain",
  t.face = "plain",
  s.face = "plain",
  l.face = "plain",
  x.angle = 0,
  y.angle = 0,
  x.hjust = 0.5,
  x.vjust = 0.5,
  y.hjust = NULL,
  y.vjust = NULL,
  legend.key.height = NULL,
  legend.key.width = NULL,
  legend.position = "right",
  legend.justification = "center"
)
}
\arguments{
\item{family}{font-family}

\item{x.size, y.size, t.size, s.size, l.size}{x axis, y axis, title, strip,
legend text size}

\item{x.face, y.face, t.face, s.face, l.face}{x axis, y axis, title, strip,
legend text face}

\item{x.angle, y.angle}{x axis, y axis text angle}

\item{x.hjust, x.vjust, y.hjust, y.vjust}{horizontal, vertical adjustment}

\item{legend.key.height, legend.key.width}{key background height & width}

\item{legend.position}{the default position of legends ("none", "left",
"right", "bottom", "top", "inside")}

\item{legend.justification}{anchor point for positioning legend inside plot
("center" or two-element numeric vector) or the justification according to
the plot area when positioned outside the plot}
}
\value{
A ggplot object
}
\description{
Simple view theme for ggplot.
}
